pyproject.toml
[tool.poetry]
name = "openfabric-test"
version = "1.1"
description = "cognitive-assistant"
authors = ["Andrei"]

[tool.poetry.dependencies]
python = "^3.8"
openfabric-pysdk = "0.2.7"
transformers = "4.36.0"
beautifulsoup4 = "4.12.2"
wikipedia = "1.4.0"
torch = {version = "^2.1.1+cpu", source = "pytorch"}
python-dotenv = "^1.0.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
